added api calls to show restaraunt and menu options
