+++ index.html
@@ -9,7 +9,7 @@
   <script src="https://js.arcgis.com/4.12/"></script>
   <script src="https://code.jquery.com/jquery-3.4.1.js" integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU=" crossorigin="anonymous"></script>
   <script src="assets/javascript/javascript.js"></script>
-  <script src="assets/javascript/arcgis.js"></script>
+  ˜
     <title>ArcGIS JavaScript Tutorials: Create a JavaScript starter app</title>
     <style>
       
@@ -33,12 +33,11 @@
       </div>
     </div>
   </div>
-
-
+  <div class="row">
+    <div class="col-12 ml-5" id="mapDiv">
     <div id="viewDiv"></div>
     <div id="mapToggle"></div>
-
-    
-    
+  </div>
+  </div>
   </body>
 </html>
